requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.86.0
anyio==4.13.0
beautifulsoup4==4.14.3
lxml
certifi==2026.2.25
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.135.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
jiter==0.13.0
markdown-it-py==4.0.0
mdurl==0.1.2
openai==2.30.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.20.0
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.0
rich==14.3.3
sniffio==1.3.1
soupsieve==2.8.3
starlette==1.0.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
requests
watchfiles==1.1.1
websockets==16.0
